move map and reserve into pages
